worker/requirements.txt
pika
psycopg2-binary
sentence-transformers
python-dotenv
scikit-learn
faiss-gpu